+++ changelog.md
@@ -61,6 +61,7 @@
      always be used.
    * Make wmllint ignore race= keys if they are part of filters inside [unit_type] (issue #4105)
    * Removed a few assserts from wmllint and postponed a few unit sanity checks to the closing of a [unit_type] tag (issue #4102)
+   * Added `tmx_trackplacer` tool, a file converter for editing map tracks with Tiled (PR #4464)
 
 ## Version 1.15.2
  ### AI: